refactor: reorder/rename fs
