+++ components/ui/draft/board/loader.tsx
@@ -1,4 +1,4 @@
-import BoardSkeleton from '@/components/skeletons/board-skeleton';
+import BoardSkeleton from '@/components/ui/draft/skeleton';
 
 const BoardLoader = () => {
    return <BoardSkeleton />;
