+++ components/ui/draft/skeleton/chat/index.tsx
