+++ components/ui/draft/skeleton/player-list/index.tsx
