+++ lib/types/index.ts
@@ -1,4 +1,4 @@
-import { Pick } from '@/components/ui/draft-order';
+import { Pick } from '@/components/ui/draft/draft-order';
 import { SupabaseClient, User } from '@supabase/supabase-js';
 import { AnchorHTMLAttributes, ReactNode } from 'react';
